really, really always shoe machine in term title
the function i'm using from OMZ for setting the title lets you override
the format for the idle title but not for a cmd, which isn't great for
long running commands (e.g. nvim). to work around this, i hook the
OMZ defined title function to tweak the parameters. at some point, it's
prolly better just to the fork the code but oh well.