don't save curdir to avoid some weirdness
if you save the current dir, auto-sessions might load a session from the
parent but then save it a child. e.g.:

/dir/subdir

if the session is restored in from /dir but, for whatever reason, the
cwd had been changed to subdir, then it'll load a session from /dir but
save one for /dir/subdir. it's possible the cwd handling would address
this but i don't currently have that enabled